requirements.txt
Pillow
customtkinter
pyperclip
numpy